// fehr/src/pages/Admin/JobPosting/types/job-posting.type.ts
export type JobPosting = {
  id: string;
  noOfVacancies: number;
  positionTitle: string;
  jobDescription: string;
  location: string;
  department: string;
  qualifications: string;
  jobCategory: string;
  employmentType: string;
  status: string;
  datePosted: string;
  deadline: string;
  salary: string;
  noOfApplicants: number;
};